Update test cases
