+++ src/index.js
@@ -4,6 +4,8 @@ import displayTodoList from './modules/displayTodo.js';
 import editTodoTask from './modules/editTask.js';
 import removeAllCheckedItem from './modules/removeAllSelected.js';
 import removeTask from './modules/removeTask.js';
+import readLocalStorage from './modules/storage.js';
+import resetIndex from './modules/resetIndex';
 
 const form = document.querySelector('.form');
 const todoInput = document.querySelector('.todo-item-input');
@@ -21,6 +23,38 @@ form.addEventListener('submit', (event) => {
   }
 });
 
+// eventListener for check/uncheck todo tasks
+//const checkboxes = document.querySelectorAll('.checkBtn');
+document.addEventListener('change', (e) => {
+  if (e.target.className === 'checkBtn') {
+    const targetElement = e.target;
+    const todoTaskTitle = targetElement.nextElementSibling.childNodes[0];
+    const todoTaskContainer = targetElement.parentElement.parentElement;
+    const { index } = todoTaskContainer.dataset;
+    const todoListList = readLocalStorage();
+  
+    if (targetElement.checked === true) {
+      todoTaskTitle.style.textDecoration = 'line-through';
+  
+      todoListList.filter((todoTask) => {
+        if (todoTask.index === index) {
+          todoTask.completed = true;
+        }
+        return false;
+      });
+    } else {
+      todoTaskTitle.style.textDecoration = 'none';
+      todoListList.filter((todoTask) => {
+        if (todoTask.index === index) {
+          todoTask.completed = false;
+        }
+        return false;
+      });
+    }
+    localStorage.setItem('todo_List', JSON.stringify(todoListList));
+  }
+});
+
 // eventListener for option and remove button
 document.addEventListener('click', (e) => {
   if (e.target.className === 'fa-solid fa-ellipsis-vertical') {
@@ -32,7 +66,7 @@ document.addEventListener('click', (e) => {
     editTodoTask(e, delBtn, optionBtn);
   } else if (e.target.className === 'fa-solid fa-trash-clock fa-trash show-trashBtn') {
     const { index } = e.target.parentElement.parentElement.dataset;
-    todoList = removeTask(index);
+    removeTask(index);
     window.location.reload();
   }
 });
@@ -41,6 +75,8 @@ document.addEventListener('click', (e) => {
 clearBtn.addEventListener('click', (e) => {
   e.preventDefault();
   removeAllCheckedItem();
+  resetIndex();
+  window.location.reload();
 });
 
 window.addEventListener('load', () => {
