+++ src/modules/removeAllSelected.js
@@ -1,9 +1,9 @@
 import readLocalStorage from './storage.js';
-import resetIndex from './resetIndex.js';
 
 const removeAllCheckedItem = () => {
+  alert(yeeeee)
   let todoList = readLocalStorage();
-
+ 
   // Remove all checked todo items
   todoList = todoList.filter((todoTask) => {
     if (todoTask.completed !== true) {
@@ -13,9 +13,7 @@ const removeAllCheckedItem = () => {
   });
   if (todoList.length > 0) {
     localStorage.setItem('todo_List', JSON.stringify(todoList));
-    resetIndex();
   }
-  window.location.reload();
 };
 
 export default removeAllCheckedItem;
